refactor: rename old directory to legacy
