+++ src/components/ui/legacy/Footer.tsx
