+++ src/components/ui/legacy/Hero.tsx
