+++ src/components/ui/legacy/Image.tsx
