+++ src/components/ui/project/Gallery.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { ChevronLeftIcon, ChevronRightIcon } from '@heroicons/react/24/outline';
-import TextHeader from '@/components/ui/old/TextHeader';
+import TextHeader from '@/components/ui/legacy/TextHeader';
 import GalleryItem from '@/components/ui/project/GalleryItem';
 import { useState } from 'react';
 import { Media, Project } from '@/types/payload-types';
