+++ src/components/ui/project/kroniipuzzle/PixiSubmissionWrapper.tsx
@@ -1,7 +1,7 @@
 import { Project } from '@/types/payload-types';
 import dynamic from 'next/dynamic';
 import fetchSubmissions, { ISubmission } from '@/lib/fetchSubmissions';
-import { getImageUrl } from '@/components/ui/old/Image';
+import { getImageUrl } from '@/components/ui/legacy/Image';
 
 interface IProps {
 	project: Omit<Project, 'flags' | 'devprops'> & {
